Working websocket connection. New post and get methods: /table/create, /table/info, /table/messaging, /user/info. Added examples to resources folder for using websocket and post/get requests by clients.

+++ src/main/java/ru/comgrid/server/api/user/UserHelp.java
@@ -2,10 +2,10 @@ package ru.comgrid.server.api.user;
 
 import org.springframework.security.oauth2.core.user.OAuth2User;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public enum UserHelp{; // no elements, utility class
-    public static BigInteger extractId(OAuth2User user){
-        return new BigInteger(((String) user.getAttributes().get("sub")));
+    public static BigDecimal extractId(OAuth2User user){
+        return new BigDecimal(((String) user.getAttributes().get("sub")));
     }
 }
